Entidad, repo Ticket

+++ BackEnd/src/main/java/com/estacsis/repository/TicketRepository.java
@@ -2,5 +2,5 @@ package com.estacsis.repository;
 import com.estacsis.entity.TicketEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<TicketEntity, Integer> {
+public interface TicketRepository extends JpaRepository<TicketEntity, Long> {
 }
